# Conflicts:
#	src/main/webapp/presentation/certainProject.jsp
#	src/main/webapp/presentation/resources/js/certainProject.js
#	src/main/webapp/presentation/resources/js/users.js

+++ src/main/java/ua/ukma/nc/dao/GroupAttachmentDao.java
@@ -18,5 +18,5 @@ public interface GroupAttachmentDao {
     
     List<GroupAttachment> getByGroupId(Long groupId);
 
-    int createGroupAttachment(GroupAttachment groupAttachment);
+    Long createGroupAttachment(GroupAttachment groupAttachment);
 }
